requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
bcrypt==4.3.0
streamlit==1.48.1
pillow==11.3.0
pandas==2.3.1
altair==5.5.0
opencv-python==4.12.0.88
ultralytics==8.3.176
